Move EditableEventCard to separate file
